+++ Beat/Export/Legacy/BeatPrint.h
@@ -8,17 +8,21 @@
 
 #import <Cocoa/Cocoa.h>
 #import "PrintView.h"
+#import "BeatEditorDelegate.h"
 
 @class Document;
 
 NS_ASSUME_NONNULL_BEGIN
 
-@class Document;
-@interface BeatPrint : NSObject <PrintViewDelegate>
-@property (weak) Document* document;
+//@class Document;
+@interface BeatPrint: NSObject <PrintViewDelegate>
+@property (weak) id<BeatEditorDelegate> document;
+//@property (weak) Document* document;
 
 - (IBAction)open:(id)sender;
 - (IBAction)openForPDF:(id)sender;
+
+- (void)loadPreview;
 
 // Move all PrintView values here
 // @property (nonatomic) NSString* header;